Move Enums to shared lib
