+++ backend/src/main/kotlin/dat257/gyro/backend/database/models/PictureRating.kt
@@ -1,7 +1,7 @@
 package dat257.gyro.backend.database.models
 
 import dat257.gyro.backend.database.tables.PictureRatings
-import dat257.gyro.backend.database.types.Ratings
+import dat257.gyro.shared.types.Rating
 import org.jetbrains.exposed.dao.IntEntity
 import org.jetbrains.exposed.dao.IntEntityClass
 import org.jetbrains.exposed.dao.id.EntityID
@@ -11,7 +11,7 @@ class PictureRating(id: EntityID<Int>) : IntEntity(id) {
 
     var profile by Profile referencedOn PictureRatings.profile
     var picture by Picture referencedOn PictureRatings.picture
-    var rating by PictureRatings.rating.transform({it.value},{ Ratings.valueOf(it.toString())})
+    var rating by PictureRatings.rating.transform({it.value},{ Rating.valueOf(it.toString())})
     var comment by PictureRatings.comment
     var createdAt by PictureRatings.createdAt
     var updatedAt by PictureRatings.updatedAt
