+++ backend/src/main/kotlin/dat257/gyro/backend/database/models/WalkPathRating.kt
@@ -1,7 +1,7 @@
 package dat257.gyro.backend.database.models
 
 import dat257.gyro.backend.database.tables.WalkPathRatings
-import dat257.gyro.backend.database.types.Ratings
+import dat257.gyro.shared.types.Rating
 import org.jetbrains.exposed.dao.IntEntity
 import org.jetbrains.exposed.dao.IntEntityClass
 import org.jetbrains.exposed.dao.id.EntityID
@@ -10,7 +10,7 @@ class WalkPathRating (id: EntityID<Int>) : IntEntity(id) {
 companion object : IntEntityClass<WalkPathRating>(WalkPathRatings)
     var profile by Profile referencedOn WalkPathRatings.profile
     var walkPath by WalkPath referencedOn WalkPathRatings.walkPath
-    var rating by WalkPathRatings.rating.transform({it.value},{ Ratings.valueOf(it.toString())})
+    var rating by WalkPathRatings.rating.transform({it.value},{ Rating.valueOf(it.toString())})
     var comment by WalkPathRatings.comment
     var createdAt by WalkPathRatings.createdAt
     var updatedAt by WalkPathRatings.updatedAt
