--- backend/src/main/kotlin/dat257/gyro/backend/database/types/Ratings.kt
@@ -1,10 +1,0 @@
-package dat257.gyro.backend.database.types
-
-enum class Ratings(val value:Int) {
-    ZERO(0),
-    ONE(1),
-    TWO(2),
-    THREE(3),
-    FOUR(4),
-    FIVE(5)
-}
