+++ shared/src/main/kotlin/dat257/gyro/shared/types/Rating.kt
@@ -1,0 +1,10 @@
+package dat257.gyro.shared.types
+
+enum class Rating(val value:Int) {
+    ZERO(0),
+    ONE(1),
+    TWO(2),
+    THREE(3),
+    FOUR(4),
+    FIVE(5)
+}
