+++ shared/src/main/kotlin/dat257/gyro/shared/types/ShowOnMap.kt
@@ -1,4 +1,4 @@
-package dat257.gyro.backend.database.types
+package dat257.gyro.shared.types
 
 enum class ShowOnMap(val value: String) {
     YES("yes"),
